pyproject.toml
[tool.poetry]
name = "aiohttp-tictactoe"
version = "0.1.0"
description = ""
authors = ["LehaDurotar <hwick1396@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
# Server-side
gunicorn = "^20.0.4"
aiohttp = "^3.7.3"
aiohttp_security = "^0.4.0"
aiohttp-session = "^2.9.0"
# Database
aiopg = "^1.1.0"
aioredis = "^1.3.1"
SQLAlchemy = "^1.3.22"
# Misc
bcrypt = "^3.2.0"
loguru = "^0.5.3"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
# Linting and formatter
black = "^20.8b1"
isort = "^5.6.4"
pytest = "^6.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line_length = 110
target_version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
length_sort = true
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 110